requirements.txt
pyautogui==0.9.54
pystray==0.19.5
Pillow==10.2.0
psutil==5.9.8
keyboard==0.13.5
pywin32==306; platform_system == "Windows" 